Add base configuration and scripts
Added docker configuration for the base environment in which the malware
is developed and tested. The environment consists of a base image
(ubuntu docker image), already infected base image (contains
whaleflu.py), and the host c2c image (contains whaleflu_c2c.py)

Additionally, updated the docker compose yaml file and README with
goals of this project and instructions on how to setup the environment.